libsel4arm-vmm: refactor virq functionality
- pull virq functionality that is not only relevant to the gicv2 into
another header file, virq.h
